requirements.txt
# ─────────────────────────────────────────
# Django Framework & REST API
# ─────────────────────────────────────────
Django>=4.2.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# API Documentation & Schema
drf-spectacular>=0.27.0

# ─────────────────────────────────────────
# Base de Données
# ─────────────────────────────────────────
psycopg2-binary>=2.9.0
dj-database-url>=2.1.0

# ─────────────────────────────────────────
# Stockage Cloudflare R2 (S3-compatible)
# ─────────────────────────────────────────
boto3>=1.34.0
django-storages[s3]>=1.14.0

# ─────────────────────────────────────────
# Fichiers statiques & WhiteNoise
# ─────────────────────────────────────────
whitenoise>=6.6.0

# ─────────────────────────────────────────
# CORS & Security
# ─────────────────────────────────────────
django-cors-headers>=4.3.0

# ─────────────────────────────────────────
# Utilitaires
# ─────────────────────────────────────────
python-dotenv>=1.0.0
Pillow>=10.0.0
PyYAML>=6.0.0
sqlparse>=0.4.0
uritemplate>=4.1.0
referencing>=0.30.0
rpds-py>=0.10.0